Pulling the XAML and Tx code back in
Portable.Xaml replaces System.Xaml. Not sure how well it works yet.
System.Transactions is not part of the .NET Standard and not a
separate library so I've added a CoreWf.Transactions library, which is
pretty much the same code as System.Transactions in .NET Core 2.
The tests are not working yet as every workflow aborts with an
exception - still a work in progress.
Also, added Microsoft.Extensions.Localization.