Added InitializeCore(null, null) back into the codepath. Allows resuming Wf at correct point after Persistance.
Thanks to @zehavibarak for the information on the missing InitializeCore.
This was part of the commented #if NET45 in the reload scenarios. (See comment // used for Load scenarios where you are rehydrating a WorkflowInstance).

Added a BookmarkConverter (new class, new file), so that we have less issues with the jsonstore.

This allows the following previously skipped tests to now pass:
UnloadFlowchartWhileExecutingFlowConditionalCondition
UnloadFlowchartWhileExecutingFlowSwitchExpression
UnloadFlowchartWhileExecutingFlowStep

The two tests which were flaky now seem to pass:
TestPersistDuringResumeBookmark
TestAbortLoad

They are timing dependent and always pass if run singly. They currently run on my RunAll.